+++ src/main/java/com/ericsson/bss/cassandra/ecaudit/auth/AuditRoleManager.java
@@ -25,7 +25,9 @@ import org.slf4j.LoggerFactory;
 
 import org.apache.cassandra.auth.AuthenticatedUser;
 import org.apache.cassandra.auth.CassandraRoleManager;
+import org.apache.cassandra.auth.IResource;
 import org.apache.cassandra.auth.IRoleManager;
+import org.apache.cassandra.auth.Permission;
 import org.apache.cassandra.auth.RoleOptions;
 import org.apache.cassandra.auth.RoleResource;
 import org.apache.cassandra.config.DatabaseDescriptor;
@@ -94,7 +96,7 @@ public class AuditRoleManager extends CassandraRoleManager
     public void createRole(AuthenticatedUser performer, RoleResource role, RoleOptions options)
             throws RequestValidationException, RequestExecutionException
     {
-        whitelistManager.verifyCreateRoleOptions(options);
+        whitelistManager.createRoleOption(options);
         super.createRole(performer, role, options);
     }
 
@@ -102,15 +104,14 @@ public class AuditRoleManager extends CassandraRoleManager
     public void alterRole(AuthenticatedUser performer, RoleResource role, RoleOptions options)
     {
         permissionChecker.checkAlterRoleAccess(performer, role, options);
-        whitelistManager.alterRoleWhitelist(performer, role, options);
+        whitelistManager.alterRoleOption(performer, role, options);
         super.alterRole(performer, role, options);
     }
 
     @Override
     public Map<String, String> getCustomOptions(RoleResource role)
     {
-        return whitelistManager.getRoleWhitelist(role).entrySet().stream()
-                               .collect(Collectors.toMap(e -> e.getKey().name(), e -> ResourceFactory.toNameCsv(e.getValue())));
+        return whitelistManager.getRoleWhitelist(role);
     }
 
     @Override