+++ src/main/java/com/ericsson/bss/cassandra/ecaudit/auth/AuditWhitelistCache.java
@@ -26,7 +26,7 @@ import org.apache.cassandra.auth.Permission;
 import org.apache.cassandra.auth.RoleResource;
 import org.apache.cassandra.config.DatabaseDescriptor;
 
-public class AuditWhitelistCache extends AuthCache<RoleResource, Map<Permission, Set<IResource>>>
+public class AuditWhitelistCache extends AuthCache<RoleResource, Map<IResource, Set<Permission>>>
 {
     private AuditWhitelistCache()
     {
@@ -63,7 +63,7 @@ public class AuditWhitelistCache extends AuthCache<RoleResource, Map<Permission,
      * @param role the Role
      * @return map of all whitelisted operation/resource combinations associated with the role
      */
-    public Map<Permission, Set<IResource>> getWhitelist(RoleResource role)
+    public Map<IResource, Set<Permission>> getWhitelist(RoleResource role)
     {
         try
         {